Code and Documentation... Complete?
So, I was able to live up to my promise in finishing the book assignment later today (rather late), but it seems as if there is much left to do in terms of making sure the actual website, and furthermore, the calculator works. I make sure to get everything ironed out and working before I can call it good on this one and move onto the final project.